move BaseGenerator to separated file
